+++ src/classes/UI5Classes/JSParser/AcornSyntaxAnalyzer.ts
@@ -103,6 +103,10 @@ export class AcornSyntaxAnalyzer {
 			if (!innerNode) {
 				innerNode = this.findAcornNode(node.params, position);
 			}
+		} else if (node.type === "Property") {
+			if (node.value) {
+				innerNode = this.findAcornNode([node.value], position);
+			}
 		} else if (node.type === "UnaryExpression") {
 			if (node.argument) {
 				innerNode = this.findAcornNode([node.argument], position);
@@ -1128,6 +1132,8 @@ export class AcornSyntaxAnalyzer {
 			if (node.left) {
 				innerNodes.push(node.left);
 			}
+		} else if (node.type === "Property") {
+			innerNodes.push(node.value);
 		} else if (node.type === "ObjectExpression") {
 			innerNodes = node.properties.map((declaration: any) => declaration.value);
 		} else if (node.type === "FunctionDeclaration" || node.type === "FunctionExpression" || node.type === "ArrowFunctionExpression") {