+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.12.76 (28-09-2021)
+* XML formatter improvements, now it formats the attribute value
+* Bugfixes for unused member linter and public member linter
+
 ## 0.12.75 (20-09-2021)
 * Method insertion position bugfixes
 
@@ -5,7 +9,7 @@
 * Bugfixes
 
 ## 0.12.73 (16-09-2021)
-* Bugfixes in determinating positions
+* Bugfixes in position determination
 
 ## 0.12.72 (07-09-2021)
 * UI5 Parser bugfix